requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
